+++ virtual_env/FINAMA/Desk/views/views_auth.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+
+def login(request):
+    return render(request, 'login.html')
+
+def register(request):
+    return render(request, 'register.html')
