+++ virtual_env/FINAMA/Desk/views/views_claim.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+# Data piutang
+def claim(request):
+    return render(request, 'data_piutang/data_piutang.html')
