+++ virtual_env/FINAMA/Desk/views/views_dashboard.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+# Dashboard with graphics
+def index(request):
+    return render(request, 'main.html')
