+++ virtual_env/FINAMA/Desk/views/views_expense.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+# Pengeluaran
+def expense(request):
+    return render(request, 'pengeluaran/pengeluaran.html')
