+++ virtual_env/FINAMA/Desk/views/views_income.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+# Pendapatan
+def income(request):
+    return render(request, 'pendapatan/pendapatan.html')
