+++ virtual_env/FINAMA/Desk/views/views_warehouse.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+# Input Barang
+def warehouse(request):
+    return render(request, 'input_barang/input_barang.html')
